requirements.txt
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
dnspython==2.2.1
ecdsa==0.18.0
exceptiongroup==1.0.1
fastapi==0.85.2
h11==0.14.0
idna==3.4
iniconfig==1.1.1
mongoengine==0.24.2
packaging==21.3
pluggy==1.0.0
prometheus-client==0.15.0
prometheus-fastapi-instrumentator==5.9.1
pyasn1==0.4.8
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
